Struct async_http_client::mpsc::UnboundedSender

pub struct UnboundedSender<T>(_);

The transmission end of a channel which is used to send values.

This is created by the unbounded method.

Methods

impl<T> UnboundedSender<T>

fn send(&mut self, msg: T) -> Result<(), SendError<T>>

Sends the provided message along this channel.

This is an unbounded sender, so this function differs from Sink::send by ensuring the return type reflects that the channel is always ready to receive messages.