Module async_http_client::prelude

Commonly needed reexports from futures and tokio-core.

Structs

Core
An event loop.

TcpStream
An I/O object representing a TCP stream connected to a remote endpoint.

Traits

Future
Trait for types which are a placeholder of a value that will become available at possible some later point in time.

Io
A trait for read/write I/O objects

Sink
A Sink is a value into which other values can be sent, asynchronously.

Stream
A stream of values, not all of which have been produced yet.